[core][refactor] remove newCommit(user, callbacks) in FileStore
